usererror checks arduino
if user sets wrong pins-> e.g. pin > MAX_UINT8

+++ src/Catty/PlayerEngine/Instructions/Arduino/ArduinoSendPWMValueBrick+Instruction.swift
@@ -31,8 +31,8 @@ extension ArduinoSendPWMValueBrick :CBInstructionProtocol {
     
     func actionBlock() -> dispatch_block_t {
         return{
-            let pinValue = UInt8(self.pin.interpretIntegerForSprite(self.script?.object))
-            let settingValue = UInt8(self.value.interpretDoubleForSprite(self.script?.object))
+            let pinValue = Int(self.pin.interpretIntegerForSprite(self.script?.object))
+            let settingValue = Int(self.value.interpretDoubleForSprite(self.script?.object))
             
             if let arduino:ArduinoDevice = BluetoothService.swiftSharedInstance.arduino {
                 arduino.setPWMArduinoPin(pinValue, value: settingValue)
